# pkg/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "contact": {
      "name": "Alyzade Shamistan",
      "email": "stalyzade@avito.ru"
    }
  },
  "tags": [
    {
      "name": "UserV1"
    }
  ],
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/v1": {
      "get": {
        "operationId": "UserV1_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_api_v1GetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserV1"
        ]
      },
      "delete": {
        "operationId": "UserV1_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserV1"
        ]
      },
      "post": {
        "operationId": "UserV1_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_api_v1CreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_api_v1CreateRequest"
            }
          }
        ],
        "tags": [
          "UserV1"
        ]
      },
      "patch": {
        "operationId": "UserV1_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_api_v1UpdateRequest"
            }
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "user_api_v1CreateRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Валидация имени: не более 50 символов"
        },
        "email": {
          "type": "string",
          "title": "Валидация email: должен быть правильным email адресом"
        },
        "password": {
          "type": "string",
          "title": "Валидация пароля: не менее 4 символов, минимум 1 заглавная буква, 1 цифра"
        },
        "passwordConfirm": {
          "type": "string",
          "title": "Валидация подтверждения пароля: должен совпадать с паролем"
        },
        "role": {
          "$ref": "#/definitions/user_api_v1Role"
        }
      }
    },
    "user_api_v1CreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "user_api_v1GetResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/user_api_v1Role"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "user_api_v1Role": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "USER",
        "ADMIN"
      ],
      "default": "UNKNOWN"
    },
    "user_api_v1UpdateRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/user_api_v1Role"
        },
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        },
        "newPasswordConfirm": {
          "type": "string"
        }
      }
    }
  }
}